news_processor/collector/pyproject.toml
[project]
name = "collector"
version = "0.1.0"
description = "Collect articles from major Vietnamese publishers and publish to Kafka."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "confluent-kafka>=2.5",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "structlog>=24.1",
    "python-dateutil>=2.9",
]

[project.scripts]
collector = "collector.app:run"

[dependency-groups]
dev = [
    "pytest>=8.2",
    "respx>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["collector"]

[tool.pytest.ini_options]
testpaths = ["tests"]
